Changed null to UNKNOWN
The nullable API might be nice to use in Kotlin but it terrible in Java because users always have to assign it to a variable and check before they can do anything.

Added additional functions to `Arch` with which it is possible to check if the architecture is of a group regardless of bitness and endianess.