Overhauled API and added many more tests
The API as it was with the unknown values was nice because there was no need to deal with `null`. However, it is also cumbersome because one is requied to check for unknown everywhere all the time. I removed it entirely and instead made everything required everywhere all the time. I also removed the `Env` class because it simply is not really useful as it is because it is impossible to determine the C ABI in a reliable way.